+++ demisto_sdk/commands/common/hook_validations/integration.py
@@ -4,7 +4,7 @@ from typing import Dict, Optional
 
 import yaml
 from demisto_sdk.commands.common.constants import (
-    BANG_COMMAND_ARGS_MAPPING_DICT, BANG_COMMAND_NAMES, DBOT_SCORES_DICT,
+    BANG_COMMAND_ARGS_MAPPING_DICT, BANG_COMMAND_NAMES, DBOT_SCORES_DICT, XSOAR_CONTEXT_STANDARD_URL,
     DEPRECATED_REGEXES, ENDPOINT_FLEXIBLE_REQUIRED_ARGS, FEED_REQUIRED_PARAMS,
     FETCH_REQUIRED_PARAMS, FIRST_FETCH, FIRST_FETCH_PARAM,
     INTEGRATION_CATEGORIES, IOC_OUTPUTS_DICT, MAX_FETCH, MAX_FETCH_PARAM,
@@ -374,7 +374,7 @@ class IntegrationValidator(ContentEntityValidator):
         Returns:
             bool. Whether a reputation command holds valid outputs
         """
-        context_standard = "https://xsoar.pan.dev/docs/integrations/context-standards"
+        context_standard = XSOAR_CONTEXT_STANDARD_URL
         commands = self.current_file.get('script', {}).get('commands', [])
         output_for_reputation_valid = True
         for command in commands:
@@ -1266,11 +1266,11 @@ class IntegrationValidator(ContentEntityValidator):
 
         if 'endpoint' not in [x.get('name') for x in commands]:
             return True
-        else:
-            # extracting the specific command from commands.
-            endpoint_command = [arg for arg in commands if arg.get('name') == 'endpoint'][0]
-            return self._is_valid_endpoint_inputs(endpoint_command, required_arguments=ENDPOINT_FLEXIBLE_REQUIRED_ARGS) \
-                and self._is_valid_endpoint_outputs(endpoint_command)
+
+        # extracting the specific command from commands.
+        endpoint_command = [arg for arg in commands if arg.get('name') == 'endpoint'][0]
+        return self._is_valid_endpoint_inputs(endpoint_command, required_arguments=ENDPOINT_FLEXIBLE_REQUIRED_ARGS) \
+            and self._is_valid_endpoint_outputs(endpoint_command)
 
     def _is_valid_endpoint_inputs(self, command_data, required_arguments):
         """
@@ -1305,17 +1305,16 @@ class IntegrationValidator(ContentEntityValidator):
         return True
 
     def _is_valid_endpoint_outputs(self, command_data):
-        context_standard = "https://xsoar.pan.dev/docs/integrations/context-standards"
+        context_standard = XSOAR_CONTEXT_STANDARD_URL
         output_for_reputation_valid = True
-        command_name = command_data.get('name')
         context_outputs_paths = set()
         for output in command_data.get('outputs', []):
             context_outputs_paths.add(output.get('contextPath'))
 
-        # validate the IOC output
-        reputation_output = IOC_OUTPUTS_DICT.get(command_name)
+        # validate the reputation command outputs
+        reputation_output = IOC_OUTPUTS_DICT.get('endpoint')
         if reputation_output and (reputation_output - context_outputs_paths):
-            error_message, error_code = Errors.missing_reputation(command_name, reputation_output,
+            error_message, error_code = Errors.missing_reputation('endpoint', reputation_output,
                                                                   context_standard)
             if self.handle_error(error_message, error_code, file_path=self.file_path):
                 self.is_valid = False