+++ CHANGELOG.md
@@ -11,8 +11,7 @@
 * Enhanced the "vX" display name validation for scripts and integrations in the **validate** command to check for every versioned script or integration, and not only v2.
 * Added the *--fail-duplicates* flag for the **create-id-set** command which will fail the command if duplicates are found.
 * Added to the **generate-docs** command automatic addition to git when a new readme file is created.
-* Added a validation in the **validate** command to ensure that '!endpoint' is configured correctly in yml file.
-
+* Added a validation in the **validate** command to ensure that the ***endpoint*** command is configured correctly in yml file.
 
 # 1.4.1
 * When in private repo without `DEMSITO_SDK_GITHUB_TOKEN` configured, get_remote_file will take files from the local origin/master.